water

+++ Assets/charger.cs
@@ -8,5 +8,5 @@ public class charger : MonoBehaviour
 	-Simply gives charge variable to individual particles
 	-To use, add this script as a component to any game object and make sure that forces.cs is running
 	*/
-    public int charge;
+    public float charge;
 }
